aotf api query playground component
