+++ apps/platform/src/pages/DiseasePage/DiseaseAssociations/DiseaseAssociations.jsx
@@ -15,9 +15,35 @@ import {
 } from "../../../components/AssociationsToolkit";
 import DISEASE_ASSOCIATIONS_QUERY from "./DiseaseAssociationsQuery.gql";
 import CopyUrlButton from "../../../components/AssociationsToolkit/CopyUrlButton";
+import ApiPlaygroundDrawer from "ui/src/components/ApiPlaygroundDrawer";
 
 function AssociationsWrapper() {
-  const { initialLoading, id } = useAotfContext();
+  const {
+    initialLoading,
+    id,
+    pageIndex,
+    pageSize,
+    searhFilter,
+    sorting,
+    enableIndirect,
+    dataSourcesWeights,
+    entity,
+    dataSourcesRequired,
+  } = useAotfContext();
+
+  const variables = {
+    id,
+    index: pageIndex,
+    size: pageSize,
+    filter: searhFilter,
+    sortBy: sorting[0].id,
+    enableIndirect,
+    datasources: dataSourcesWeights,
+    entity,
+    aggregationFilters: dataSourcesRequired,
+  };
+
+  console.log(DISEASE_ASSOCIATIONS_QUERY);
 
   if (initialLoading) return <AotFLoader />;
 
@@ -37,6 +63,10 @@ function AssociationsWrapper() {
           </OptionsControlls>
         </Box>
         <Box>
+          <ApiPlaygroundDrawer
+            query={DISEASE_ASSOCIATIONS_QUERY.loc.source.body}
+            variables={variables}
+          />
           <TargetPrioritisationSwitch />
         </Box>
       </ControlsSection>
